project/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
langchain
langchain-community
langchain_openai
openai
azure-search-documents
python-docx
PyPDF2
azure-identity